Sort the examples in the CF template
The code previously produced CF code from a listing of files from the
fs.readdirSync function. The problem is that the order of the files
returned by fs.readdirSync is determined by the underlying filesystem
implementation, and thus changes from one filesystem to the next.

This was causing the template code to be produced with elements in a
different order from one developer to the next. This did not result
in any functional difference (or bug) in the template, but it did
cause the template tests to fail because the the snapshot verification
expects a fixed order.