Allow deployments using SSO profiles
The previous logic verified that a profile exists by looking for the
configuration value `aws_access_key_id`. This value only exists for
for profiles with API keys, so it fails for SSO authenticated profiles.

This commit uses the more generic `aws configure list` command to
verify that the profile exists. The command returns 0 if the profile
is present, and 255 if it is not.